# t8ira-swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Codepeak Track 8 Child API Base",
        "description": "This is the REST API for Track 8 Child applications",
        "termsOfService": "",
        "contact": {
            "name": "Codepeak Technologies Pvt. Ltd.",
            "url": "https://codepeaktechnologies.com/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Development Server"
        },
        {
            "url": "https://cpdaily-api.onrender.com",
            "description": "Production Server"
        }
    ],
...
    "paths": {
...
        "/review": {
            "post": {
                "summary": "Post Review API for Interns",
                "description": "Posting a daily review by intern",
                "tags": [
                    "Review"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/reviewSave"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Review successfully"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Review exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 409
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Review exists"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/byintern": {
            "get": {
                "summary": "List of all previous reviews by an intern.",
                "tags": [
                    "Review"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "tutor": {
                                                        "type": "object",
                                                        "properties": {
                                                            "id": {
                                                                "type": "string",
                                                                "description": "The id of the tutor data",
                                                                "example": "6726317cmnmxznc128738127"
                                                            },
                                                            "userid": {
                                                                "type": "string",
                                                                "description": "The userid of the tutor data",
                                                                "example": "6726317cmnmxznc128738127"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "description": "The name of the tutor",
                                                                "example": "Rainu Bennan"
                                                            },
                                                            "phone": {
                                                                "type": "string",
                                                                "description": "the phone number of the tutor",
                                                                "example": "9999999999"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "description": "the emailid of the tutor",
                                                                "example": "name`@`mail.com"
                                                            },
                                                            "qualification": {
                                                                "type": "string",
                                                                "description": "the qualification of the tutor",
                                                                "example": "BCA"
                                                            }
                                                        }
                                                    },
                                                    "year": {
                                                        "type": "number",
                                                        "description": "The year of the selected date.",
                                                        "example": 2023
                                                    },
                                                    "month": {
                                                        "type": "number",
                                                        "description": "The month of the selected date.",
                                                        "example": 1
                                                    },
                                                    "day": {
                                                        "type": "number",
                                                        "description": "The day of the selected date.",
                                                        "example": 1
                                                    },
                                                    "comment": {
                                                        "type": "string",
                                                        "description": "The comment by the intern.",
                                                        "example": "Today's session was good."
                                                    },
                                                    "rating": {
                                                        "type": "number",
                                                        "description": "The rating by the intern",
                                                        "example": 4.5
                                                    },
                                                    "createdby": {
                                                        "type": "string",
                                                        "description": "The id of the user who created the entry",
                                                        "example": "6726317cmnmxznc128738127"
                                                    },
                                                    "createdat": {
                                                        "type": "string",
                                                        "description": "The time when the entry was made",
                                                        "example": "[datetime]"
                                                    },
                                                    "updatedby": {
                                                        "type": "string",
                                                        "description": "The id of the user who editted the entry",
                                                        "example": "6726317cmnmxznc128738127"
                                                    },
                                                    "updatedat": {
                                                        "type": "string",
                                                        "description": "The time when the entry was editted",
                                                        "example": "[datetime]"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "number",
                                            "description": "the api call status",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "the api message",
                                            "example": "Server error"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "reviewSave": {
                "type": "object",
                "properties": {
                    "tutorid": {
                        "type": "string",
                        "description": "The selected Tutor Id",
                        "example": "6726317cmnmxznc128738127"
                    },
                    "year": {
                        "type": "number",
                        "description": "The year of the selected date.",
                        "example": 2023
                    },
                    "month": {
                        "type": "number",
                        "description": "The month of the selected date.",
                        "example": 1
                    },
                    "day": {
                        "type": "number",
                        "description": "The day of the selected date.",
                        "example": 1
                    },
                    "comment": {
                        "type": "string",
                        "description": "The comment by the intern.",
                        "example": "Today's session was good."
                    },
                    "rating": {
                        "type": "number",
                        "description": "The year of the selected date.",
                        "example": 4.5
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}